MC-32923: Remove legacy logic from migration data class and refactor new code
