+++ etc/commerce-to-commerce/map-eav.xml.dist
@@ -150,18 +150,6 @@
                 </handler>
             </transform>
             <transform>
-                <field>eav_attribute_set.attribute_set_name</field>
-                <handler class="\Migration\Handler\AddPrefix">
-                    <param name="prefix" value="Migration_" />
-                </handler>
-            </transform>
-            <transform>
-                <field>eav_attribute_group.attribute_group_name</field>
-                <handler class="\Migration\Handler\AddPrefix">
-                    <param name="prefix" value="Migration_" />
-                </handler>
-            </transform>
-            <transform>
                 <field>eav_attribute_group.default_id</field>
                 <handler class="\Migration\Handler\SetValue">
                     <param name="value" value="0" />
