+++ etc/opensource-to-commerce/map-eav.xml.dist
@@ -149,18 +149,6 @@
                     <param name="value" value="100" />
                 </handler>
             </transform>
-            <transform>
-                <field>eav_attribute_set.attribute_set_name</field>
-                <handler class="\Migration\Handler\AddPrefix">
-                    <param name="prefix" value="Migration_" />
-                </handler>
-            </transform>
-            <transform>
-                <field>eav_attribute_group.attribute_group_name</field>
-                <handler class="\Migration\Handler\AddPrefix">
-                    <param name="prefix" value="Migration_" />
-                </handler>
-            </transform>
             <ignore>
                 <field>catalog_eav_attribute.is_configurable</field>
             </ignore>
