+++ etc/opensource-to-opensource/map-eav.xml.dist
@@ -148,19 +148,6 @@
                 </handler>
             </transform>
             <transform>
-                <field>eav_attribute_set.attribute_set_name</field>
-                <handler class="\Migration\Handler\AddPrefix">
-                    <param name="prefix" value="Migration_" />
-                </handler>
-            </transform>
-
-            <transform>
-                <field>eav_attribute_group.attribute_group_name</field>
-                <handler class="\Migration\Handler\AddPrefix">
-                    <param name="prefix" value="Migration_" />
-                </handler>
-            </transform>
-            <transform>
                 <field>catalog_eav_attribute.apply_to</field>
                 <handler class="\Migration\Handler\EavAttribute\ApplyTo" />
             </transform>
