+++ src/Migration/Step/Eav/Helper.php
@@ -181,100 +181,6 @@ class Helper
     }
 
     /**
-     * Get added groups
-     *
-     * @return \Migration\ResourceModel\Record[]
-     */
-    public function getAddedGroups()
-    {
-        return $this->addedGroups;
-    }
-
-    /**
-     * Set added groups
-     *
-     * @param array $addedGroups
-     * @return void
-     */
-    public function setAddedGroups(array $addedGroups)
-    {
-        $this->addedGroups = $addedGroups;
-    }
-
-    /**
-     * Get design attribute and groups data
-     *
-     * @return array
-     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
-     */
-    public function getDesignAttributeAndGroupsData()
-    {
-        $scheduleGroupsMigrated = [];
-        $catalogProductSetIdsMigrated = [];
-        $catalogProductSetIdDefault = null;
-        $customLayoutAttributeId = null;
-        $customDesignAttributeId = null;
-        $entityTypeIdCatalogProduct = $this->getDestinationRecords('eav_entity_type', ['entity_type_code'])
-            ['catalog_product']['entity_type_id'];
-
-        foreach ($this->getDestinationRecords('eav_attribute_set') as $record) {
-            if ($entityTypeIdCatalogProduct == $record['entity_type_id']) {
-                if ('Default' == $record['attribute_set_name']) {
-                    $catalogProductSetIdDefault = $record['attribute_set_id'];
-                } else {
-                    $catalogProductSetIdsMigrated[] = $record['attribute_set_id'];
-                }
-            }
-        }
-        foreach ($this->getDestinationRecords('eav_attribute_group') as $group) {
-            if ('schedule-design-update' == $group['attribute_group_code'] &&
-                $catalogProductSetIdDefault != $group['attribute_set_id']
-            ) {
-                $scheduleGroupsMigrated[] = $group;
-            }
-        }
-        foreach ($this->getDestinationRecords('eav_attribute') as $record) {
-            if ($record['entity_type_id'] == $entityTypeIdCatalogProduct) {
-                switch ($record['attribute_code']) {
-                    case 'custom_layout':
-                        $customLayoutAttributeId = $record['attribute_id'];
-                        break;
-                    case 'custom_design':
-                        $customDesignAttributeId = $record['attribute_id'];
-                        break;
-                }
-            }
-        }
-
-        return [
-            'scheduleGroupsMigrated' => $scheduleGroupsMigrated,
-            'catalogProductSetIdsMigrated' => $catalogProductSetIdsMigrated,
-            'customLayoutAttributeId' => $customLayoutAttributeId,
-            'customDesignAttributeId' => $customDesignAttributeId,
-            'entityTypeIdCatalogProduct' => $entityTypeIdCatalogProduct
-        ];
-    }
-
-    /**
-     * Remove ignored attributes from source records
-     *
-     * @param array $sourceRecords
-     * @return array
-     */
-    public function clearIgnoredAttributes($sourceRecords)
-    {
-        $ignoredAttributes = $this->getAttributesGroupCodes('ignore');
-        foreach ($sourceRecords as $attrNum => $sourceAttribute) {
-            if (in_array($sourceAttribute['attribute_code'], array_keys($ignoredAttributes))
-                && in_array($sourceAttribute['entity_type_id'], $ignoredAttributes[$sourceAttribute['attribute_code']])
-            ) {
-                unset($sourceRecords[$attrNum]);
-            }
-        }
-        return $sourceRecords;
-    }
-
-    /**
      * Retrieves attribute codes with types of $groupName group
      *
      * @param string $groupName
